feat: Profiling 检索页面掉帧优化 --Story=121833786 # Reviewed, transaction id: 30317
