+++ bkmonitor/webpack/src/trace/pages/main/span-details.tsx
@@ -37,7 +37,7 @@ import { copyText, deepClone, random } from 'monitor-common/utils/utils';
 import ExceptionGuide, { type IGuideInfo } from '../../components/exception-guide/exception-guide';
 import MonitorTab from '../../components/monitor-tab/monitor-tab';
 import { formatDate, formatDuration, formatTime } from '../../components/trace-view/utils/date';
-import ProfilingFlameGraph from '../../plugins/charts/profiling-graph/flame-graph/flame-graph';
+import ProfilingFlameGraph from '../../plugins/charts/profiling-graph/profiling-flame-graph/flame-graph';
 import FlexDashboardPanel from '../../plugins/components/flex-dashboard-panel';
 import { useIsEnabledProfilingInject } from '../../plugins/hooks';
 import { BookMarkModel } from '../../plugins/typings';
